added REST endpoint for tier info
